refactor(db): Split commits_fetcher.rs into directory module
